+++ ebcli/lib/codecommit.py
@@ -101,15 +101,14 @@ def list_branches(repo_name, next_token=None):
 
 
 def region_supported(region):
-    supported_regions = ["us-east-1", "us-east-2"]
+    supported_regions = ["us-east-1", "us-east-2", "us-west-2", "eu-west-1"]
     if region is not None and region in supported_regions:
         return True
 
     return False
 
 
-def _sign_codecommit_url(region, url_to_sign):
-    credentials = aws.get_credentials()
+def _sign_codecommit_url(credentials, region, url_to_sign):
     signer = SigV4Auth(credentials, 'codecommit', region)
     request = AWSRequest()
     request.url = url_to_sign
@@ -135,7 +134,7 @@ def _sign_codecommit_url(region, url_to_sign):
 def create_signed_url(remote_url):
     split_url = remote_url.split("//")
     credentials = aws.get_credentials()
-    password = _sign_codecommit_url(aws.get_region_name(), remote_url)
+    password = _sign_codecommit_url(credentials, aws.get_region_name(), remote_url)
     username = credentials.access_key
 
     if credentials.token is not None: