+++ ebcli/resources/strings.py
@@ -203,7 +203,10 @@ strings = {
     'appversion.cannotdeploy': 'Some application versions failed to process. Unable to continue deployment.',
     'appversion.processtimeout': 'All application versions have not reached a "Processed" state. Unable to continue with deployment.',
     'instance.processes.health': '{healthy}/{total} processes healthy.',
+
+    # codesource
     'codesource.info': 'Configures the code source for the EB CLI to use by default.',
+    'codesource.localmsg': 'Default set to use local sources',
 }
 
 prompts = {
@@ -274,7 +277,10 @@ prompts = {
                                         'Create an IAM role called "{}" with appropriate permissions to continue, or specify a role with --service-role.\n'
                                         'See http://docs.aws.amazon.com/elasticbeanstalk/latest/dg/concepts-roles.html for more info. \nActual error: {}',
     'general.pressenter': 'Press enter to continue',
-    'compose.groupname': 'Please enter the group name to be used'
+    'compose.groupname': 'Please enter the group name to be used',
+
+    # codesource
+    'codesource.codesourceprompt': 'Select your codesource',
 }
 
 alerts = {
@@ -328,6 +334,7 @@ flag_text = {
     'create.group': 'group suffix',
     'create.modules': 'a list of modules',
     'create.elb_type': 'load balancer type',
+    'create.source': 'source of code to create from directly; example source_location/repo/branch',
 
     # Deploy
     'deploy.env': 'environment name',
@@ -338,7 +345,7 @@ flag_text = {
     'deploy.nohang': 'return immediately, do not wait for deploy to be completed',
     'deploy.staged': 'deploy files staged in git rather than the HEAD commit',
     'deploy.group_suffix': 'group suffix',
-    'deploy.source': 'source of code to deploy directly; example repo/branch',
+    'deploy.source': 'source of code to deploy directly; example source_location/repo/branch',
 
     # Events
     'events.follow': 'wait and continue to print events as they come',
@@ -349,7 +356,7 @@ flag_text = {
     'init.keyname': 'default EC2 key name',
     'init.interactive': 'force interactive mode',
     'init.module': 'module directory',
-    'init.source': 'source of code to set as default; example repo/branch',
+    'init.source': 'source of code to set as default; example source_location/repo/branch',
 
     # labs
     'labs.cwl.remove': 'remove .ebextensions',
@@ -397,7 +404,7 @@ flag_text = {
 
     # use
     'use.env': 'environment name',
-    'use.source': 'source of code to set as default; example repo/branch',
+    'use.source': 'source of code to set as default; example source_location/repo/branch',
     'use.repo': 'default code commit repository',
     'use.branch': 'default code commit branch will use default repository if none is specified',