+++ CHANGES.rst
@@ -3,6 +3,12 @@ Changelog
 =========
 
 ------------------
+3.8.4 (2016-11-16)
+------------------
+- Added '--source' flag to create, deploy, init and use to directly use source from CodeCommit repositories
+- Added us-west-2 (Oregon) and eu-west-1 (Ireland) CodeCommit support
+
+------------------
 3.8.3 (2016-10-17)
 ------------------
 - Added us-east-2 (Ohio) CodeCommit support